refactor: consolidate local learning writes
