+++ docs/agents/architecture.md
@@ -34,7 +34,7 @@ Before changing background commands or sync execution, read [ADR-0001](../adr/00
 [ADR-0003](../adr/0003-single-document-learning-data.md) records the single-document design. The document is authoritative at runtime. The superseded migration lifecycle and collection/settings persistence wrappers have been removed; historical compatibility lives in the startup bridge and pure document conversions.
 
 - One browser-local document owns its schema version, modification timestamp, slug-keyed cards with embedded notes, daily statistics, and stored settings overrides. Card timestamps use epoch milliseconds; `dataUpdatedAt` is an optional timestamp string, and statistics use local `YYYY-MM-DD` dates. Missing settings retain normal defaults, including browser-language detection.
-- Domain owns current schemas and explicit-input policy. Named local workflows prepare complete changes and pass their captured edit time to `services/save-local-learning-document.ts`. This shared save path prepares the timestamp and calls the writer, which owns final completed-document validation before one write. Command payload schemas own early input validation and settings-update normalization; workflows consume those parsed inputs without repeating those checks. Reviews commit cards, statistics, and their edit timestamp together; queue calculations use one captured document and time. Keep normal operations behind startup readiness and the existing background write queue.
+- Domain owns current schemas and explicit-input policy. Named local workflows in `services/learning.ts` prepare complete changes and pass their captured edit time to a private shared save helper. This shared save path prepares the timestamp and calls the writer, which owns final completed-document validation before one write. Command payload schemas own early input validation and settings-update normalization; workflows consume those parsed inputs without repeating those checks. Reviews commit cards, statistics, and their edit timestamp together; queue calculations use one captured document and time. Keep normal operations behind startup readiness and the existing background write queue.
 - `infrastructure/storage/learning-document.ts` owns current-document reads, the shared required-document check, and validating replacement through the shared storage keys. `learning-document-conversions.ts` owns backup parsing and pure version-conversion orchestration shared by startup, import, and Gist pull. `document-conversions/` owns the frozen historical contracts and individual conversions. Preserve supported historical versions and their meaning, frozen historical data contracts, and precedence rules. Validate historical input before stripping extras and validate the current result. Keep browser storage, clocks, per-step loaders/savers/cleanup, physical-layout modules, and intermediate completion writes out of the conversion sequence.
 - One startup bridge gathers scattered legacy values only when the new document is absent. Prepare and validate before writing the complete document and its version. After success the document is authoritative; cleanup failure must not trigger re-import, and corrupt or future documents must not fall back to stale legacy keys. Preserve credentials and already-combined connection values over retained legacy fields.
 - File backups and Gist payloads use the document representation. Import validates and converts before one replacement, preserves the imported timestamp, and clears omitted notes/settings through replacement. Retain legacy backup timestamp fallback without treating conversion as a learner edit.
