+++ entrypoints/popup/queries/__tests__/settings.test.tsx
@@ -8,7 +8,7 @@ import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { sendMessage } from '@/infrastructure/browser/messages';
 import { readLearningDocument, replaceLearningDocument } from '@/infrastructure/storage/learning-document';
 import { getSettings } from '@/infrastructure/storage/learning-queries';
-import { updateSettings } from '@/services/settings';
+import { updateSettings } from '@/services/learning';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
 import { cardQueryKeys } from '../cards';
